+++ client/components/Glass/index.ts
@@ -1,0 +1,3 @@
+import { Glass as _Glass } from './Glass';
+
+export const Glass = _Glass();
