+++ client/components/index.ts
@@ -1,5 +1,5 @@
 export * from './Atoms';
-export * from './Glass/Glass';
+export * from './Glass';
 export * from './Inspector';
 export * from './NullComp/NullComp';
 export * from './Navigation';
